Move homepage logic to util.service
